Added context api
